Beginnings of tone parameter support
For the sake of updates to this project, and to test my ability to commit to someone else's repo, began work on attempting to to fill up the Tone tab in the VST via showing controls for the wave group and the waveform selection per tone (no names for now: only numbers). The two parameters are not editable at this time due to needing to find the time to implement a whole automatic working SysEx request for tones, but if nothing else, it can currently provide some good reference for sound design back in the day.