Intermediate relaxing of ROMs required
While the main repo requires all ROM's and their correct checksums in place, this commit will only nag about missing the required five. In addition, added a slot for a custom SR-JV80 ROM of one's choice, intended for testing custom ROM images using my tools in a separate repo, thus, it is spared of checksum checking as well.

It would be ideal to load the optional expansion ROM's dynamically, and with more programming knowledge, I could try later but this is a good intermediate step imho.