Partial support of drumkit editing (#72)
* Parial support of drumkit editing

Feature support may be getting closer to completion thanks to this commit. Most parameters in the rhythms sets are surprisingly similar to their patch equivalents, though also thoroughly tested the Mute Group parameter before committing. I don't expect the implementation for the other exclusive parameter (Envelope Mode) to be perfect due to so far failing to find rhythm tones that use it, so may correct later if collaborators don't beat me to it, but feel free to test all the same!

---------

Co-authored-by: EvilDragon <mario.kruselj@gmail.com>